- added disconnect() method
- reorganized unit tests and added test for disconnect
