+++ build.gradle
@@ -4,7 +4,7 @@ apply plugin: 'signing'
 apply plugin: 'maven'
 
 group = 'com.keysolutions'
-version = '0.5.7.2'
+version = '0.5.7.3'
 
 // get rid of bootclasspath warning by setting it
 def env = System.getenv()
@@ -36,7 +36,7 @@ repositories {
 }
 
 dependencies {
-    compile group: 'commons-collections', name: 'commons-collections', version: '3.2'
+    compile group: 'commons-collections', name: 'commons-collections', version: '3.2.+'
     testCompile group: 'junit', name: 'junit', version: '4.+'
     compile group: 'org.java-websocket', name: 'Java-WebSocket', version: '1.3.0'
     compile group: 'com.google.code.gson', name: 'gson', version: '2.2.4'
